requirements.txt
openai
transformers
huggingface_hub
pandas
python-dotenv
openpyxl
torch
